Move :timeslice to TimeComputations.scala
