+++ core/src/main/scala/filodb.core/metadata/SimpleComputations.scala
@@ -82,44 +82,6 @@ object SimpleComputations {
   }
 
   /**
-   * Syntax: :timeslice <colName> <durationString>
-   * Valid for: LongColumn, TimestampColumn
-   * Bucketizes the time value into a time bucket by the <durationString> which has a format like
-   * "10s" - 10 seconds, "2m" - 2 minutes, "5h" - 5 hours... this must be a time string recognized
-   * by the Typesafe Config library.
-   * Produces a Long column with the bucketed time in milliseconds.
-   */
-  object TimesliceComputation extends SingleColumnComputation {
-    def funcName: String = "timeslice"
-
-    def parseDurationMillis(arg: String): Long Or InvalidComputedColumnSpec = {
-      try {
-        val config = ConfigFactory.parseString(s"a = $arg")
-        Good(config.getMilliseconds("a"))
-      } catch {
-        case e: Exception => Bad(BadArgument(e.getMessage))
-      }
-    }
-
-    def analyze(expr: String,
-                dataset: String,
-                schema: Seq[Column]): ComputedColumn Or InvalidComputedColumnSpec = {
-      for { info <- parse(expr, schema, Set(LongColumn, TimestampColumn))
-            duration <- parseDurationMillis(info.param) }
-      yield {
-        val func = (info.colType match {
-          case LongColumn =>
-            (l: Long) => l / duration * duration
-          case TimestampColumn =>
-            (t: Timestamp) => t.getTime / duration * duration
-          case o: Column.ColumnType => ???
-        }).asInstanceOf[info.keyType.T => Long]
-        computedColumnWithDefault(expr, dataset, info, LongColumn, LongKeyType)(-1L)(func)
-      }
-    }
-  }
-
-  /**
    * :stringPrefix stringCol numChars
    * Defaults to "" if null string column
    */
